feat: Implement local caching for custom movie lists
This commit introduces local caching for custom movie lists and their memberships.

Key changes include:
- Added Room entities: `MovieListLocal`, `MovieListMembershipLocal`, and `SyncMetadataLocal`.
- Implemented DAOs: `MovieListDao`, `ListMembershipDao`, and `SyncMetadataDao`.
- Created `CustomMovieListLocalDataSource` and its implementation to manage local data.
- Updated `CustomMovieListRepositoryImpl` to integrate with the local data source, including cache refresh logic and mutex for synchronization.
- Introduced `MovieListSyncWorker` to periodically refresh the cache in the background using WorkManager.
- Configured `NovixApplication` to initialize WorkManager and schedule the periodic sync.
- Updated Hilt modules to provide new DAOs, local data source, and WorkManager.
- Modified `id` and `moviesCount` in `MovieList` entity from `UInt` to `Int`.
- Added necessary dependencies for WorkManager and Hilt.

+++ data/src/main/java/com/london/data/mapper/list/customListMapper.kt
@@ -5,7 +5,7 @@ import com.london.data.utils.orZero
 import com.london.domain.entity.MovieList
 
 fun CustomMovieListResponse.toEntity(): MovieList = MovieList(
-    id = this.id.orZero().toUInt(),
+    id = this.id.orZero(),
     name = this.name.orEmpty(),
-    moviesCount = this.itemCount.orZero().toUInt()
+    moviesCount = this.itemCount.orZero()
 )
